requirements.txt
fastapi
uvicorn
absl-py==2.1.0
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.1
gTTS==2.5.2
idna==3.7
jax==0.4.30
jaxlib==0.4.30
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pycparser==2.22
pygame==2.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sounddevice==0.4.7
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
